+++ src/shared.ts
@@ -1,3 +1,0 @@
-const validWordBoundaries = new Set('  []()-–—\'"""'.split("").concat([".", ",", ":", ";", "/"]));
-
-export const isValidWordBoundary = (char: string): boolean => validWordBoundaries.has(char);